requirements.txt
discord.py
youtube_dl
pytube
flask